+++ src/config/feeds-schema.ts
@@ -1,3 +1,6 @@
+import {CalculatorName} from '../types/Calculator.js';
+import {FetcherName} from '../types/fetchers.js';
+
 export default {
   description: 'Feeds schema for the JSONSchema validator',
   type: 'object',
@@ -42,6 +45,7 @@ export default {
             {$ref: '#/definitions/CoinmarketcapHistoHourFetcher'},
             {$ref: '#/definitions/CoinmarketcapHistoDayFetcher'},
             {$ref: '#/definitions/OnChainDataFetcher'},
+            {$ref: '#/definitions/UniswapV3Fetcher'},
             {$ref: '#/definitions/OptionsPriceFetcher'},
             {$ref: '#/definitions/YearnVaultTokenPriceFetcher'},
             {$ref: '#/definitions/RandomNumberFetcher'},
@@ -49,6 +53,7 @@ export default {
             {$ref: '#/definitions/GoldApiPriceFetcher'},
             {$ref: '#/definitions/MetalPriceApiFetcher'},
             {$ref: '#/definitions/MetalsDevApiPriceFetcher'},
+            {$ref: '#/definitions/SovrynPriceFetcher'},
           ],
         },
         calculator: {
@@ -66,7 +71,7 @@ export default {
     },
     GVolImpliedVolatilityFetcher: {
       properties: {
-        name: {const: 'GVolImpliedVolatility'},
+        name: {const: FetcherName.G_VOL_IMPLIED_VOLATILITY},
         params: {
           type: 'object',
           properties: {
@@ -82,7 +87,7 @@ export default {
     },
     CryptoComparePriceFetcher: {
       properties: {
-        name: {const: 'CryptoComparePrice'},
+        name: {const: FetcherName.CRYPTO_COMPARE_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -98,7 +103,7 @@ export default {
     },
     CryptoComparePriceWSFetcher: {
       properties: {
-        name: {const: 'CryptoComparePriceWS'},
+        name: {const: FetcherName.CRYPTO_COMPARE_PRICE_WS},
         params: {
           type: 'object',
           properties: {
@@ -115,7 +120,7 @@ export default {
     },
     CryptoCompareHistoHourFetcher: {
       properties: {
-        name: {const: 'CryptoCompareHistoHour'},
+        name: {const: FetcherName.CRYPTO_COMPARE_HISTO_HOUR},
         params: {
           type: 'object',
           properties: {
@@ -132,7 +137,7 @@ export default {
     },
     CryptoCompareHistoDayFetcher: {
       properties: {
-        name: {const: 'CryptoCompareHistoDay'},
+        name: {const: FetcherName.CRYPTO_COMPARE_HISTO_DAY},
         params: {
           type: 'object',
           properties: {
@@ -149,7 +154,7 @@ export default {
     },
     CoinmarketcapPriceFetcher: {
       properties: {
-        name: {const: 'CoinmarketcapPrice'},
+        name: {const: FetcherName.COINMARKETCAP_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -165,7 +170,7 @@ export default {
     },
     CoinmarketcapHistoHourFetcher: {
       properties: {
-        name: {const: 'CoinmarketcapHistoHour'},
+        name: {const: FetcherName.COINMARKETCAP_HISTO_HOUR},
         params: {
           type: 'object',
           properties: {
@@ -182,7 +187,7 @@ export default {
     },
     CoinmarketcapHistoDayFetcher: {
       properties: {
-        name: {const: 'CoinmarketcapHistoDay'},
+        name: {const: FetcherName.COINMARKETCAP_HISTO_DAY},
         params: {
           type: 'object',
           properties: {
@@ -199,7 +204,7 @@ export default {
     },
     CoingeckoPriceFetcher: {
       properties: {
-        name: {const: 'CoingeckoPrice'},
+        name: {const: FetcherName.COINGECKO_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -215,7 +220,7 @@ export default {
     },
     PolygonIOPriceFetcher: {
       properties: {
-        name: {const: 'PolygonIOPrice'},
+        name: {const: FetcherName.POLYGON_IO_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -230,7 +235,7 @@ export default {
     },
     PolygonIOStockPriceFetcher: {
       properties: {
-        name: {const: 'PolygonIOStockPrice'},
+        name: {const: FetcherName.POLYGON_IO_STOCK_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -245,7 +250,7 @@ export default {
     },
     PolygonIOCryptoPriceFetcher: {
       properties: {
-        name: {const: 'PolygonIOCryptoPrice'},
+        name: {const: FetcherName.POLYGON_IO_CRYPTO_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -261,7 +266,7 @@ export default {
     },
     IEXEnergyFetcher: {
       properties: {
-        name: {const: 'IEXEnergy'},
+        name: {const: FetcherName.IEX_ENERGY},
         params: {
           type: 'object',
           properties: {
@@ -276,7 +281,7 @@ export default {
     },
     BEACPIAverageFetcher: {
       properties: {
-        name: {const: 'BEACPIAverage'},
+        name: {const: FetcherName.BEACPI_AVERAGE},
         params: {
           type: 'object',
           properties: {
@@ -291,7 +296,7 @@ export default {
     },
     TWAPGasPriceFetcher: {
       properties: {
-        name: {const: 'TWAPGasPrice'},
+        name: {const: FetcherName.TWAP_GAS_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -307,7 +312,7 @@ export default {
     },
     OnChainDataFetcher: {
       properties: {
-        name: {const: 'OnChainData'},
+        name: {const: FetcherName.ON_CHAIN_DATA},
         params: {
           type: 'object',
           properties: {
@@ -327,9 +332,26 @@ export default {
       required: ['params'],
       additionalProperties: false,
     },
+    UniswapV3Fetcher: {
+      properties: {
+        name: {const: FetcherName.UNISWAP_V3},
+        params: {
+          type: 'object',
+          properties: {
+            fromChain: {type: 'array'},
+            token0: {type: 'string'},
+            token1: {type: 'string'},
+          },
+          required: ['fromChain', 'token0', 'token1'],
+          additionalProperties: false,
+        },
+      },
+      required: ['params'],
+      additionalProperties: false,
+    },
     GoldApiPriceFetcher: {
       properties: {
-        name: {const: 'GoldApiPrice'},
+        name: {const: FetcherName.GOLD_API_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -345,7 +367,7 @@ export default {
     },
     MetalPriceApiFetcher: {
       properties: {
-        name: {const: 'MetalPriceApi'},
+        name: {const: FetcherName.METAL_PRICE_API},
         params: {
           type: 'object',
           properties: {
@@ -361,7 +383,7 @@ export default {
     },
     MetalsDevApiPriceFetcher: {
       properties: {
-        name: {const: 'MetalsDevApi'},
+        name: {const: FetcherName.METALS_DEV_API},
         params: {
           type: 'object',
           properties: {
@@ -377,7 +399,7 @@ export default {
     },
     PolygonIOCurrencySnapshotGramsFetcher: {
       properties: {
-        name: {const: 'PolygonIOCurrencySnapshotGrams'},
+        name: {const: FetcherName.POLYGON_IO_CURRENCY_SNAPSHOT_GRAMS},
         params: {
           type: 'object',
           properties: {
@@ -392,12 +414,12 @@ export default {
     },
     OptionsPriceFetcher: {
       properties: {
-        name: {const: 'OptionsPrice'},
+        name: {const: FetcherName.OPTIONS_PRICE},
       },
     },
     YearnVaultTokenPriceFetcher: {
       properties: {
-        name: {const: 'YearnVaultTokenPrice'},
+        name: {const: FetcherName.YEARN_VAULT_TOKEN_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -413,7 +435,7 @@ export default {
     },
     RandomNumberFetcher: {
       properties: {
-        name: {const: 'RandomNumber'},
+        name: {const: FetcherName.RANDOM_NUMBER},
         params: {
           type: 'object',
           properties: {
@@ -423,9 +445,27 @@ export default {
       },
       additionalProperties: false,
     },
+    SovrynPriceFetcher: {
+      properties: {
+        name: {const: FetcherName.SOVRYN_PRICE},
+        params: {
+          type: 'object',
+          properties: {
+            base: {type: 'string'},
+            quote: {type: 'string'},
+            amountInDecimals: {type: 'number'},
+            quoteDecimals: {type: 'number'},
+          },
+          required: ['base', 'quote', 'amountInDecimals'],
+          additionalProperties: false,
+        },
+      },
+      required: ['params'],
+      additionalProperties: false,
+    },
     YearnTransformPriceCalculator: {
       properties: {
-        name: {const: 'YearnTransformPrice'},
+        name: {const: CalculatorName.YEARN_TRANSFORM_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -440,7 +480,7 @@ export default {
     },
     OptionsPriceCalculator: {
       properties: {
-        name: {const: 'OptionsPrice'},
+        name: {const: CalculatorName.OPTIONS_PRICE},
         params: {
           type: 'object',
           properties: {
@@ -455,19 +495,19 @@ export default {
     },
     IdentityCalculator: {
       properties: {
-        name: {const: 'Identity'},
+        name: {const: CalculatorName.IDENTITY},
       },
       additionalProperties: false,
     },
     TWAPCalculator: {
       properties: {
-        name: {const: 'TWAP'},
+        name: {const: CalculatorName.TWAP},
       },
       additionalProperties: false,
     },
     VWAPCalculator: {
       properties: {
-        name: {const: 'VWAP'},
+        name: {const: CalculatorName.VWAP},
       },
       additionalProperties: false,
     },